+++ src/constants/errors/index.js
@@ -9,4 +9,6 @@ module.exports = {
   SAFLE_ID_NOT_REGISTERED: 'This SafleID is not registered.',
   NO_SAFLEID: 'No SafleID is registered for this index number.',
   INVALID_ENV_INPUT: 'Invalid env input.',
+  COIN_ADDRESS_NOT_REGISTERED: 'This coin address is not registered.',
+  NO_COIN_ADDRESS: 'No coin address registered.',
 };